Fix stale Excel export tests
Align compute and Lakebase expectations with current pricing behavior, and split DBSQL VM coverage to preserve the suite's file-size guard.
